requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30

# AI (OpenRouter — uses OpenAI-compatible SDK)
openai==1.30.1
tiktoken==0.7.0

# Google
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0

# Jobs / scheduling
apscheduler==3.10.4

# Media
python-multipart==0.0.9
Pillow==10.3.0
pydub==0.25.1

# Auth / security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
tenacity==8.3.0
pytz==2024.1

python-telegram-bot==21.x